# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "API Documentation",
    "description": "RESTful API Documentation - User, Role, Category Management System",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/api",
...
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "in": "header",
      "name": "Authorization",
      "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\""
    }
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/categories/": {
      "get": {
        "tags": [
          "Categories"
        ],
        "summary": "Get all categories",
        "description": "Retrieve a list of all categories",
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Categories retrieved successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/categories/add": {
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "Add a new category",
        "description": "Create a new category",
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Category information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Electronics"
                },
                "is_active": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category created successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/categories/update": {
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "Update an existing category",
        "description": "Update category information",
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Category update information",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "507f1f77bcf86cd799439014"
                },
                "name": {
                  "type": "string",
                  "example": "Home Electronics"
                },
                "is_active": {
                  "type": "boolean",
                  "example": true
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category updated successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/categories/delete": {
      "post": {
        "tags": [
          "Categories"
        ],
        "summary": "Delete a category",
        "description": "Delete a category from the system",
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "description": "Category ID to delete",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "_id": {
                  "type": "string",
                  "example": "507f1f77bcf86cd799439014"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category deleted successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/routes/": {
      "post": {
        "tags": [
          "Upload"
        ],
        "summary": "Upload a file",
        "description": "Upload a single file to the server",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "accept-language",
            "in": "header",
            "type": "string"
          },
          {
            "name": "file",
            "in": "formData",
            "type": "file",
            "required": true,
            "description": "File to upload"
          }
        ],
        "responses": {
          "200": {
            "description": "File uploaded successfully"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
...
}